refactor : 팀 퇴장시 분석 결과 업데이트 로직을 TeamResultAnalysisService로 이주
