+++ aiku/aiku-main/src/main/java/aiku_main/service/team/TeamResultAnalysisService.java
@@ -8,6 +8,7 @@ import aiku_main.dto.team.result.racing_odds.TeamRacingResult;
 import aiku_main.dto.team.result.racing_odds.TeamRacingResultDto;
 import aiku_main.repository.team.TeamRepository;
 import common.domain.team.Team;
+import common.domain.team.TeamResult;
 import common.util.ObjectMapperUtil;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -50,5 +51,62 @@ public class TeamResultAnalysisService {
         TeamRacingResultDto resultDto = new TeamRacingResultDto(team.getId(), result);
 
         team.setTeamRacingResult(ObjectMapperUtil.toJson(resultDto));
+    }
+
+    @Transactional
+    public void updateTeamResultOfExitMember(Long memberId, Long teamId) {
+        Team team = teamRepository.findTeamWithResult(teamId).orElseThrow();
+
+        updateLateTimeResultOfExitMember(memberId, team);
+        updateBettingTimeResultOfExitMember(memberId, team);
+        updateRacingTimeResultOfExitMember(memberId, team);
+    }
+
+    private void updateLateTimeResultOfExitMember(Long memberId, Team team){
+        TeamResult teamResult = team.getTeamResult();
+        if (teamResult.hasNoLateTimeResult()) {
+            return;
+        }
+
+        TeamLateTimeResultDto result = ObjectMapperUtil.parseJson(teamResult.getLateTimeResult(), TeamLateTimeResultDto.class);
+        result.getMembers().forEach(resultMember -> {
+            if (resultMember.getMemberId().equals(memberId)) {
+                resultMember.setTeamMember(false);
+            }
+        });
+
+        team.setTeamLateResult(ObjectMapperUtil.toJson(result));
+    }
+
+    private void updateBettingTimeResultOfExitMember(Long memberId, Team team){
+        TeamResult teamResult = team.getTeamResult();
+        if (teamResult.hasNoTeamBettingResult()) {
+            return;
+        }
+
+        TeamBettingResultDto result = ObjectMapperUtil.parseJson(teamResult.getTeamBettingResult(), TeamBettingResultDto.class);
+        result.getMembers().forEach(resultMember -> {
+            if (resultMember.getMemberId().equals(memberId)) {
+                resultMember.setTeamMember(false);
+            }
+        });
+
+        team.setTeamBettingResult(ObjectMapperUtil.toJson(result));
+    }
+
+    private void updateRacingTimeResultOfExitMember(Long memberId, Team team){
+        TeamResult teamResult = team.getTeamResult();
+        if (teamResult.hasNoTeamRacingResult()) {
+            return;
+        }
+
+        TeamRacingResultDto result = ObjectMapperUtil.parseJson(teamResult.getTeamRacingResult(), TeamRacingResultDto.class);
+        result.getMembers().forEach(resultMember -> {
+            if (resultMember.getMemberId().equals(memberId)) {
+                resultMember.setTeamMember(false);
+            }
+        });
+
+        team.setTeamRacingResult(ObjectMapperUtil.toJson(result));
     }
 }
