+++ aiku/aiku-main/src/main/java/aiku_main/service/team/TeamService.java
@@ -3,20 +3,14 @@ package aiku_main.service.team;
 import aiku_main.dto.team.*;
 import aiku_main.application_event.publisher.TeamEventPublisher;
 import aiku_main.dto.*;
-import aiku_main.dto.team.result.betting_odds.TeamBettingResultDto;
-import aiku_main.dto.team.result.late_time.TeamLateTimeResultDto;
-import aiku_main.dto.team.result.racing_odds.TeamRacingResult;
-import aiku_main.dto.team.result.racing_odds.TeamRacingResultDto;
 import aiku_main.exception.MemberNotFoundException;
 import aiku_main.exception.TeamException;
 import aiku_main.repository.member.MemberRepository;
 import aiku_main.repository.schedule.ScheduleRepository;
 import aiku_main.repository.team.TeamRepository;
-import common.util.ObjectMapperUtil;
 import common.domain.member.Member;
 import common.domain.team.Team;
 import common.domain.team.TeamMember;
-import common.domain.team.TeamResult;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
@@ -37,7 +31,6 @@ public class TeamService {
     private final ScheduleRepository scheduleRepository;
     private final MemberRepository memberRepository;
     private final TeamEventPublisher teamEventPublisher;
-    private final ObjectMapperUtil objectMapperUtil;
 
     @Transactional
     public Long addTeam(Long memberId, TeamAddDto teamDto){
@@ -114,66 +107,6 @@ public class TeamService {
         checkTeamMember(memberId, teamId, true);
 
         return team.getTeamResult() == null? null : team.getTeamResult().getTeamRacingResult();
-    }
-
-    @Transactional
-    public void updateTeamResultOfExitMember(Long memberId, Long teamId) {
-        Team team = findTeamWithResult(teamId);
-        if (team.getTeamResult() == null) {
-            return;
-        }
-
-        updateLateTimeResultOfExitMember(memberId, team);
-        updateBettingTimeResultOfExitMember(memberId, team);
-        updateRacingTimeResultOfExitMember(memberId, team);
-    }
-
-    private void updateLateTimeResultOfExitMember(Long memberId, Team team){
-        TeamResult teamResult = team.getTeamResult();
-        if (teamResult.getLateTimeResult() == null) {
-            return;
-        }
-
-        TeamLateTimeResultDto result = objectMapperUtil.parseJson(teamResult.getLateTimeResult(), TeamLateTimeResultDto.class);
-        result.getMembers().forEach(resultMember -> {
-            if (resultMember.getMemberId().equals(memberId)) {
-                resultMember.setTeamMember(false);
-            }
-        });
-
-        team.setTeamLateResult(objectMapperUtil.toJson(result));
-    }
-
-    private void updateBettingTimeResultOfExitMember(Long memberId, Team team){
-        TeamResult teamResult = team.getTeamResult();
-        if (teamResult.getTeamBettingResult() == null) {
-            return;
-        }
-
-        TeamBettingResultDto result = objectMapperUtil.parseJson(teamResult.getTeamBettingResult(), TeamBettingResultDto.class);
-        result.getMembers().forEach(resultMember -> {
-            if (resultMember.getMemberId().equals(memberId)) {
-                resultMember.setTeamMember(false);
-            }
-        });
-
-        team.setTeamBettingResult(objectMapperUtil.toJson(result));
-    }
-
-    private void updateRacingTimeResultOfExitMember(Long memberId, Team team){
-        TeamResult teamResult = team.getTeamResult();
-        if (teamResult.getTeamRacingResult() == null) {
-            return;
-        }
-
-        TeamRacingResultDto result = objectMapperUtil.parseJson(teamResult.getTeamRacingResult(), TeamRacingResultDto.class);
-        result.getMembers().forEach(resultMember -> {
-            if (resultMember.getMemberId().equals(memberId)) {
-                resultMember.setTeamMember(false);
-            }
-        });
-
-        team.setTeamRacingResult(objectMapperUtil.toJson(result));
     }
 
     private Member findMember(Long memberId){
